refactor(tests): split test_pr_command.py into a focused pr_command package (#443)
`tests/teatree_core/test_pr_command.py` was 1221 LOC and the largest
remaining standalone oversized test module after #896
(`management_commands/`), #912 (`sync/`), #913 (`utils/`) and #914
(`models/`). It packed every `pr` subcommand plus the command's
private helpers into one file. This splits it into a focused
`tests/teatree_core/pr_command/` package, one module per concern.

Pure verbatim reorganization: every class and the `_shippable_ticket`
helper and `clear_overlay_cache` fixture are AST-identical to the
original (verified with `ast.dump`); only import paths changed and the
module-level autouse `clear_overlay_cache` fixture moved verbatim into a
package `conftest.py` (mirroring #912). Zero production code touched,
zero assertions changed.

Relates-to #443
